:art: refactored to move some constants to global
