+++ frontend/src/components/employees/employees.module.scss
@@ -15,7 +15,6 @@ h1 {
 
 .addButton {
   grid-area: addbutton;
-  margin: 2em;
 }
 
 Table {
@@ -33,18 +32,14 @@ Table {
   border: none;
   outline: none;
   background-color: transparent;
-  font-weight: 100;
   margin: 0.15em;
-  font-size: 1em;
+  &:hover {
+    font-size: 1.2em;
+    margin: 0em;
+  }
 }
 
 .removeButton {
   color: red;
   font-weight: 1000;
-}
-
-.editButton:hover,
-.removeButton:hover {
-  font-size: 1.2em;
-  margin: 0em;
 }
